Updated test accuracy assertion check as well
Updated test accuracy assertion check as well (earlier updated params and loss values to assert against since there was a bug in the sigmoid cross entropy loss calculation. Once that bug was fixed, the test case also needed to be updated to assert against the right param and loss values).