Make torch optional
Torch is a huge dependency it pulls in a ~700MB wheel from pypi and when
extracted it takes multiple gigabytes of disk space. It is only directly
used in the neural_network module and since it doesn't work on windows
we alread have a check to see if it's installed or not and handle that
gracefully. Instead of requiring it be installed and every aqua install
we should make this optional with a setuptools extras_require. So you
can just run: pip install qiskit-aqua[neural-network] to install torch
if you are using that component.