cannot use `del` on numpy arrays + comment why +l offset is dropped
del is replaced with numpy.delete. since delete is only there as a 
safeguard and seldomly called it's preferable to implement as call to 
delete instead of e.g. having a list of valid indices, since latter 
always requires access to the array. or at least it needs an additional 
if-clause to test if access is necessary.