Make pyeda optional
Installing pyeda requires a compiler on all environments because they
only publish an sdist and no pre-compiled binaries. This requires every
user who installs aqua (and by extensions the qiskit meta-repository) to
have a working compiler setup. It also slows down the installation
process because the installer has to compile the code for pyeda. The use
of pyeda is already optional because it is not pip installable on
windows so we're properly handling the case where it's not installed
already. This commit makes the dependency optional and adds it as a
setuptools extra to make installing it if it's needed easier.

Fixes #551