This commit contains an unfinished checkpoint for modifying BinaryTree.preparation_circuit to allow for optional control on another register.
This commit is correct except for the case of a single qubit state with negative amplitudes with a control register.

The best way to fix this issue is to change how sign information is handled for a single qubit.

This could break the current implementation (which is correct with no control), hence the commit.

Revert to this commit if all hell breaks loose.