Modified BinaryTree.preparation_circuit to:
(1) Modify input circuit in place.
(2) Allow for a variable number of registers to prepare the state in.

The main reason for this was that preparing an eigenstate of the unitary W for QSVE must be done
across two registers (the row and col registers).

Refactored unit tests and QSVE class for this change.