Remove jsonschema cap
We have had the jsonschema version capped since it was capped in terra.
However there was no clear reason for this in terra and it limits
interoperability with other python packages. Especially because
jsonschema is commonly used library across the python ecosystem. The
version we were capping to was also quite old being release 2.5 years
ago. This commit removes the cap so that people can run aqua in
environments where newer jsonschema is required.