Add pypi automation for releases
As part of streamlining the release process, this commit adds automation
for publishing to pypi. For the aqua this is straightforward since it is
pure python and we just need to build an sdist and a python3 only wheel
and then upload both to pypi. This takes another manual step out of the
release process and all that is necessary is to push a tag (or use
github's ui for creating a tag and/or release). The pypi upload will be
handled automatically from that action.