Remove terra cap from stable branch
Qiskit terra 0.10.0 was just released and there is not a corresponding
feature release for aqua we need to remove the cap. With the cap in
place everytime someone manually installed aqua with newer terra
installed pip will downgrade it. When we push the meta package release
in Qiskit/qiskit#635 pip will report an error in bold red that we're
violating the aqua requirement. This commit removes the version cap on
the stable branch. Nothing in terra 0.10.0 has broken aqua and this
avoids the error message from pip.