Add python 3.8 support and deprecate python 3.5
This commit adds support for running aqua under python 3.8. It does
this by adding the trove classifiers to the package metadata and adding
a test job for python 3.8 At the same time this commit starts the
deprecation window for python 3.5. Python 3.5 goes end of life by
upstream python in Sept. 2020. We should give our users running with
python 3.5 (which is about 10% of qiskit users this year based on pypi
data) fair notice that when upstream python stops supporting it we do as
well. Corresponding to this change, terra is going to start raising a
warning when anything from qiskit.* is imported, see:
Qiskit/qiskit-terra#3268