Add tox.ini to start using tox
This commit adds support for using tox by adding a tox.ini. This makes
it very simple for users and contributors to run tests and style checks
locally by just using the `tox` command. It will handle building and
installing aqua and it's python dependencies in a virtualenv to isolate
it from system python for testing. This is also a common tool for python
projects, as many (if not most) other python projects use it. Most of
the other qiskit elements also support using it too.

In my local environment pyenchant was not installable to avoid similar
issues on other systems this commit removes it from the requirements-dev
list and only tries to install it in the lint tox job. This prevents it
from trying to be installed except when users try to run lint, since
it's not a requirement for unit testing.