Remove reno lint from lint command
Reno lint is only used if a project is using reno for release notes
(which qiskit-terra does). Aqua isn't using them to handle it's release
documentation yet. This will likely change in the future, but for right
now there is no reason to run reno lint as part of the lint job.