Update tox.ini to use stestr too
In #665 we migrated the default test runner used for running aqua's unit
tests to be stestr to get the benefits of parallel execution. However
tox support merged around the same time and did not take this into
account. So right now the test runner used between 'make test' and tox
are different. This commit updats the tox job definitions to use stestr
too instead of the stdlib unittest runner.