fix AEBase getters
The idea of the a_factory setter is to allow:
```
ae = AmplitudeEstimation(m)
for a_factory in all_my_factories:
  ae.a_factory = a_factory
  result = ae.run(quantum_instance)
```
however, with the current implementation of `check_factories` and getter of `q_factory`,
the `q_factory` gets set in the first iteration of the for loop and then re-used, which is of course wrong.
If the `a_factory` changes, so does q.

This commit fixes this issue, by always returning the `q_factory` corresponding to the `a_factory`, unless `q_factory` was explicitely set.