bug fix "referenced before assignment"
at the line 865 rise the error:

" UnboundLocalError: local variable 'qc' referenced before assignment"

when "state_in" and "quantum_registers" are provided. 
This is due to the fact that when the if at line 863 is satisfied and the error at line 865 doesn't arise no "qc" variable is declared, generating the error.