Cleanup travis config
This commit makes a few small changes to simplify the travis
configuration. The first thing it does is decouple the pypi upload job
from the common test setup config. There is no reason to build aer and
terra from source or install all the dependencies explicitly as part of
the release workflow. Building and sdist and wheel will take care of any
dependencies that need to be installed as part of the build process, run
time dependencies are not needed. Then it changes the pypi job version
to 3.7 while it's being modified just so we don't have to worry about it
later when 3.6 goes EoL in 2021. The last thing done here is that with
the release of stestr 2.6.0 in Decemeber the workaround for
https://github.com/mtreinish/stestr/issues/266 is no longer needed
because the issue has been fixed.