Check for num_atoms in QMolecule.core_orbitals()
In the example of a `freeze_core` Hamiltonian this method is called to
return the list of core orbitals.
When working with the FCIDumpDriver it may be the case that `num_atoms`
is not set. Thus, we check for this case and return in empty list
instead.

Unittests have been added which assert that a warning is logged.