Adds OperatorBase support in aux_operators setters
aux_ops setters in vqe and numpy_eigen_solver call to_opflow on all operators,
whereas this conversion is only necessary for LegacyBaseOperator ---
BaseOperator does not need it and does not support it.

This commit adds a check in both setters, applying the conversion only
to instances of LegacyBaseOperator.

Fixes #1018