Remove cvxopt and use cvxpy instead
This commit removes the usage of cvxopt from aqua. It was supposed to be
removed as part of the 0.7 release since some users have issue
installing it, and other users aren't allowed to have GPL code installed
in their environment. However, cvxopt is still used for the quadratic
program solver utility. cvxopt is not needed to solve a quadratic
program and other elements are already doing this with cvxpy which
doesn't have the same issues [1][2]. This commit migrates the cvxopt
usage to cvxpy to be consistent with the rest of qiskit.

[1] https://github.com/Qiskit/qiskit-aer/commit/e41d8683775d1d905def9e91048d8e338bc221fa
[2] https://github.com/Qiskit/qiskit-ignis/commit/42cbbf8821a7d920ac44620d1e09ab48300f4be8