Fix hash function of PauliOp class
This change replaces the use of the `id()` function in the `PauliOp`
hash function with the hash of its string representation. This is inline
with the hash function of the `Pauli` class.

This fix is necessary, since prior to this change, a set comparison
including instances of identical `PauliOp`s would fail, because sets
rely on hash tables.

A regression test to ensure the working state after this commit is
included, too.

Credits for spotting and collaborating in fixing the bug go to @Cryoris