Make cvxpy an optional dependency
This commit makes the cvxpy dependency optional because many users have
trouble installing it because on many environments it requires that a
compiler be present to build the package from sdist and there are no
precompiled binary wheels present for many environments. Users of the
svm classifier will have to install cvxpy manually or with the extra
now, but it should remove the barrier for most users.

Related to #1053