Set qcp to True for cvxpy Problem.solve in qp_solver
The legacy tutorial in the Qiskit/qiskit-tutorials repo for running
QSVM classification is failing with newer versions of aqua that use
cvxpy. The error message returned suggests setting the qcp flag to true
on solve(). This treats the problem as disciplined quasiconvex program
instead of a disciplined convex program and seems to avoid the hard
failure in the tutorial.

Fixes: Qiskit/qiskit-tutorials#928