Make pyscf an optional dependency
pyscf doesn't package precompiled binaries for windows, older versions
of macOS or 32bit linux and this has been a constant source of pain for
users trying to install the qiskit meta-package. The windows case was
already being handled so pyscf isn't a hard requirement for aqua. This
commit removes pyscf from the requirements list and makes it an optional
dependency for all environments.

Fixes Qiskit/qiskit#696