Remove cvxpy cap
The bug in cvxpy 1.1.x that was breaking the aqua usage has been fixed
in the latest 1.1.3 release (see Qiskit/qiskit-aer#832 and
Qiskit/qiskit-aer#779 for more details). Since the bug is fixed we can
remove the version cap an enable newer versions of cvxpy to be
installed. Exclude rules are left on the setuptools extra requirement so
that end users don't accidently try to use a version known not to run.