Fix incorrect cycle detection in VQEAdapt
Fixes #1184

The previous regex used for cyclicity detection had the problem of not
restricting itself to checking whole numbers. Thus, it would detect a
cycle even if only single digits of the indices repeated (e.g. as in
[190, 126, 62] where the 6 "repeats").
This bug is fixed by ensuring that the regex will only detect sequences
of at least two whole numbers in a row.

Since this change would allow single numbers to repeat indefinitely, the
case of single repeating indices is handled explicitly for the last two
entries. Due to the iterative nature of the algorithm this suffices to
catch all intermediate steps, too.

In order to allow the simple unittesting of this functionality, it has
been extracted into a separate method of the VQEAdapt class.