Fix cycle regex to allow sequences longer than 2
The previous introduction of the word boundaries around the numbers of
the first regex group resulted in the regex being unable to detect
sequences of any length other than two.
That was the also the initial reason for using ".+" rather than "\d+"
which allows multiple numbers to be matched. Due to the regex greedy
matching "(.+ .+)" will now truely try match as many numbers as possible
(starting from at least two).

The word boundaries around "\1" still ensure that we are correctly
detecting a pattern of whole numbers and don't split them into digits.