Add non-iterative phase estimation algorithms
This adds an implemenation of the non-iterative (multi-qubit phase
register) QPE algorithm. There are two layers. In the lower layer,
the user supplies a unitary. In the higher layer, the user supplies
a Hermitian operator (maybe a Hamiltonian). In the latter case,
facilities for scaling and exponentiating the operator to generate
the unitary are provided. The same structures that do the rescaling
have methods to extract the eigenvalues of the original Hamiltonian.
The intention is for the user to have a range of choices for how much
of the algorithm to control manually and how much is done automatically.