Implement the AdaptVQEGroundStateCalculation class
This properly implements the AdaptVQEGroundStateCalculation class.
The idea of this class is to allow easy use of the VQEAdapt algorithm
which was previously extended to support the MinimumEigensolver
interface. This Calculation class is solely for the use with this
algorithm which is why it does not support setting a different solver.
Instead, it requires setting a QuantumInstance object through its
initializer.

The changes which this requires include the following:

- fixing the GroundStateCalculation class
- fixing the MinEigensolverGroundStateCalculation class

Furthermore, a very basic unittest for this change was added.