Switch bksf functions to use retworkx (#1261)
* Switch bksf stabilizers function to use retworkx

This commit migrates the only remaining networkx usage in aqua to use
retworkx. The recent 0.5.0 release added the necessary functionality for
the bksf stabilizer function. Local testing shows that there is not
really a performance difference between networkx and retworkx in this
case since the run time of the function is dominated by the generating
the edge list and creating the operator objects, but it does enable us to
remove the networkx dependency since there is nothing else using it in
aqua.

* Migrate vacuum_operator() too

* Fix typo

* Fix another typo

* Update docplex tests to use retworkx

* Add networkx as a dev requirement

* Use retworkx to generate adjacency matrix in TestDocplex