Port VQEAdapt to AdaptVQE (based on a GroundStateCalculation)
This properly refactors the old
qiskit.chemistry.algorithms.minimum_eigen_solver.VQEAdapt into the new
qiskit.chemistry.ground_state_calculation.AdaptVQE class which
implements the new GroundStateCalculation interface.

The name change is intentional because the new name is in-line with the
literature on AdaptVQE. The old name was presumably intentional to
highlight its relation to VQE.

The unittests for VQEAdapt are ported accordingly.