Remove check for real coefficients in CircuitStateFn.from_vector
Before this commit, a ValueError is raised if operators.CircuitStateFn.from_vector
is passed a vector with any non-positive component. There is a comment stating
that this is neccessary because qiskit.extensions.Initializer also has this
requirement. But, this is apparently not the case, as is stated explicitly
in the documentation for Initializer.... Perhaps there was such a
restriction earlier which has since been removed. But, it doesn't seem likely.
The reason for the check is not clear.

Closes #1276