[WIP] basic Result-interfaces
This adds the following interfaces:
- qiskit.chemistry.ChemistryResult(AlgorithmResult)
- qiskit.chemistry.ground_state_calculation.GroundStateResult(ChemistryResult)
- qiskit.chemistry.ground_state_calculation.FermionicGroundStateResult(GroundStateResult)

The logic is as follows:
ChemistryResult and GroundStateResult are empty interfaces which define
a hierarchy for the result classes. They provide a useful granularity
for type checking. Once more subclasses are added we can extract common
functionality up into the stack as needed.

The FermionicGroundStateResult currently is just a port of the old
MolecularChemistryResult and MolecularGroundStateResult. Both of these
classes are deprecated in this commit. Since this class is currently
specific to the GroundStateCalculation it is part of that module. This
is, however, subject to change.