[wip] resolve cyclic import problems
Having a `Result` class in either the `ground_state_calculations` or
`qubit_operator_transformations` modules leads to a coupling of the two
modules that is too tight. I.e. we run into cyclic imports whenever we
try to run any code because the `QubitOpTransformation` class needs to
load the `Result` class while at the same time it needs to be available
in the `GroundStateCalculation` class. Thus, when an additional coupling
in between these two modules is created, we have cyclic import.

Thus, we drop one level of granularity and only provide a single
result-interface on the `qiskit.chemistry` level, here, called
`ChemistryResult` (open for discussion). I also moved the `DipoleTuple`
to this level to simplify its usage in the other classes.