Filter DeprecationWarnings in unittests
This filters all of the DeprecationWarnigns introduced in this PR.
In 3 months from now we can remove these tests once the corresponding
code is also removed.

This commit also adds TODO labels to all of these unittest files which
will have to be migrated to work within the new framework.