Fix Trotter evolution of sums with identity terms
Previously, Trotterization did not exclude any terms consisting of a scalar times the
identity from the list of terms to be repeated. Instead, if present, an identity term was
included in the resulting circuit as repeated identity gates, with the coefficient of the
identity operator discarded.  But, this coefficient must be accounted for: exp(-i * coeff
* I) should introduce a global phase.

With this PR, we remove the scalar-multiple-of-identity terms from the SummedOp to be
Trotterized. We then set the global phase of the resulting circuit to account for the
removed identity term.

Fixes #1269