Rethink the qiskit.chemistry.results
After some discussion we decided to rethink parts of the
qiskit.chemistry.results interfaces. We want to minimize the
differentiation between grround and excited states and let only the
concrete `QubitOperatorTransformation` handle the fermionic and bosonic
differentiation.

Thus, I apply the following changes:
* provide a single `EigenstateResult` which can be either a ground or
  excited state
* derive `ElectronicStructureResult` (prev. `FermionicResult`) from this
* use the `FermionicTransformation.interpret()` method to map a general
  `EigenstateResult` to the type specific for the transformation

With the last change we also align the `qiskit.chemistry` module further
with the `qiskit.optimization` module where we also use `interpret()`
rather than `add_context()` methods (which was the name we intended to
use prior to this commit).