Add an `evaluate_operators` method to the GroundStateCalculation
This method can be used to evaluate any additional auxiliary operators
after the GSC has finished. This will be necessary especially in the
excited state calculations and is in general a nice option for the user
to have.

The method is similar to how the `VQE` evaluates auxiliary operators
internally. However, it can also deal with `MinimumEigensolver`s which
do not use a `QuantumInstance`.
A curiousity arose because the `VQE` appears to wrap the auxiliary
operator results into lists multiple times. In order to ensure
interoperability the same is done here.

The `AdaptVQE` now derives off of the `MinimumEigensolverGSC` because
there was no obvious drawback to this and the benefit of reusing this
new method was welcome.

Unittests have been added to test the new functionality.