Update test_grad.py
Several high-level function calls still used a `method=...` argument in the initialization; however, we changed this to `grad_method`, `hess_method`, `qfi_method`. 
So I updated these arguments. 

This was causing some tests to not be carried out as the default argument was inferred when the correct argument name was not assigned a method type.