Implement fix for 1q MatrixEvolution
HamiltonionPE fails in general with MatrixEvolution. This commit
decomposes (twice) the unitary returned by convert. This allows
HamiltonianPE to give correct eigenvalues for some Hamiltonians
in combination with MatrixEvolution. Namely, all 1Q Hamiltonians
that I tested now work. A test is included. It is rather crude in
order that it be fast.