Do not use the default_filter_criterion for excited states
Since we opted to make the default filter criterion of the fermionic
case filter non-singlet states this kind of misses the point in the
excirted states cases.
Thus we now provide a custom filter in the test.

For the backlog we should consider a better implementation of the
default filter which can handle this automatically.