Ensure aux_operator eigenvalues are normalized
As reported on multiple occasions (#1460 and #1467), the eigenvalues of
the auxiliary operators are not correctly normalized when the QASM
backend is used.

This commit fixes this short-coming by applying the same normalization
to the VQE's eigenstate when obtained from a QASM backend (in which case
this is a dictionary) as done by the `CircuitSampler` in its
`sample_circuits` function.
The case of the statevector backend is unaffected by this change, as it
will return the eigenstate as a list.

Co-authored-by: Julien Gacon <gaconju@gmail.com>