Add support for Python 3.9 (#1504)
* Add support for Python 3.9

Python 3.9.0 was released on 10-05-2020, this commits marks the support
of Python 3.9 in qiskit-aqua. It adds the supported python version in
the package metadata and updates the CI configuration to run test jobs
on Python 3.9. This follows on from Qiskit/qiskit-terra#5189,
Qiskit/qiskit-ignis#505, and Qiskit/qiskit-aer#1071

* Add release note

* Remove 3.9 tutorial job

Jupyter on 3.9 doesn't have a good story yet, so just run tutorials on
3.6 and 3.8 for now.

* Update optional requirement system constraints

Co-authored-by: Steve Wood <40241007+woodsp-ibm@users.noreply.github.com>