Stop using basic aer internal private attributes
This commit fixes the usage of internal only explicitly private
attributes of basicaer's job class. These look they were used to
manually create a job id which was manually set ahead of time despite
still being a random uuid and then launching a simulation from that
constructed job object. Which is both the incorrect order to run a
simulation (runs generate jobs, not the other way around) but also
served no functional purpose. This commit removes this as this is a
blocker for Qiskit/qiskit-terra#5218 which is upgrading the basic aer
provider to use the latest versioned provider interface which removes
all of those private methods and also moves to a synchronous execution
model.